Add component option to prompt package
Introduced 'component' parameter to prompt package which lets the user specify whether to use an "input" field or a "textarea" field. An error is thrown if an invalid component type is entered. Also added a condition that is the 'Enter' key to trigger the resolver only if the component is an "input". Updated the package version in package.json.